requirements.txt
accelerate==1.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.1
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyvips==3.0.0
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
timm==1.0.20
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing-extensions==4.15.0
ultralytics==8.3.218
ultralytics-thop==2.0.17
urllib3==2.5.0
